% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphidocs-package.R
\docType{data}
\name{colours}
\alias{colours}
\title{Some useful colours}
\format{
An object of class \code{character} of length 11.
}
\usage{
colours
}
\description{
Some useful colours
}
\keyword{datasets}
