% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_oracle.R
\name{consult_oracle}
\alias{consult_oracle}
\title{Consult the Pythia}
\usage{
consult_oracle(query)
}
\arguments{
\item{query}{something you would like to know}
}
\value{
a random response from the oracle
}
\description{
Need some help? Ask the oracle.
}
\examples{
consult_oracle(12345L)
consult_oracle("A")
}
