% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delphidocs-package.R
\docType{package}
\name{delphidocs-package}
\alias{delphidocs}
\alias{delphidocs-package}
\title{delphidocs: pkgdown template for Delphi}
\description{
A pkgdown template for the CMU Delphi R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmu-delphi/delphidocs}
  \item Report bugs at \url{https://github.com/cmu-delphi/delphidocs/issues}
}

}
\author{
\strong{Maintainer}: Daniel J. McDonald \email{daniel@stat.ubc.ca}

Other contributors:
\itemize{
  \item Delphi Research Group [copyright holder, funder]
}

}
