% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\docType{data}
\name{responses}
\alias{responses}
\title{Some things the oracle might say}
\format{
An object of class \code{character} of length 19.
}
\usage{
responses
}
\description{
Some things the oracle might say
}
\keyword{datasets}
